removed nearly all use of std::cout/cerr for the new NGLMessage sub-system (NGLMessage itself still uses std::cerr

+++ include/ngl/Obj.h
@@ -89,9 +89,9 @@ class NGL_DLLEXPORT Obj : public AbstractMesh
     // f v/vt/vn v/vt/vn v/vt/vn v/vt/vn
     bool parseFaceVertexNormalUV(std::vector<std::string> &_tokens) noexcept;
     // as faces can use negative index values keep track of index
-    size_t m_currentVertexOffset=0;
-    size_t m_currentNormalOffset=0;
-    size_t m_currentUVOffset=0;
+    int m_currentVertexOffset=0;
+    int m_currentNormalOffset=0;
+    int m_currentUVOffset=0;
     bool m_isLoaded=false;
 };
